server/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
click==8.1.7
colorlog==6.8.2
dill==0.3.8
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
isort==5.13.2
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==1.26.4
platformdirs==4.2.0
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pylint==3.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
